+++ fe/src/pages/Login/index.tsx
@@ -1,12 +1,15 @@
 import { Link } from 'react-router-dom';
 
-import LoginForm from '@/components/LoginForm';
+import LoginForm from './LoginForm';
+import * as S from './stlye';
 
 export default function Login() {
   return (
     <div>
       <h1>ISSUE TRACKER</h1>
-      <button>GitHub 계정으로 로그인</button>
+      <S.GithubLoginButton onClick={() => {}} disabled={false}>
+        GitHub 계정으로 로그인
+      </S.GithubLoginButton>
       <div>or</div>
       <LoginForm />
       <Link to="/signup">회원가입</Link>
